lint file Input
